requirements.txt
opencv_python==4.13.0.90
pillow==12.1.0
ultralytics==8.4.6
